requirements.txt
playwright
pandas
google-generativeai
langchain
langgraph
chromadb
python-dotenv
streamlit
sentence-transformers